requirements.txt
fastapi==0.104.1
python-dotenv==1.0.0
redis_om==0.2.1
uvicorn==0.24.0.post1